added filters module, and resolved some tasks after my code review
