+++ js/main.js
@@ -1,17 +1,24 @@
 import { openBigPhotoPopup } from './big-picture-popup.js';
-import { getPictures } from './api.js';
+import { getData } from './api.js';
 import { closeUploadOverlay, setUploadFormSubmit } from './upload-form.js';
 import { renderPictures, showDataError } from './render-pictures.js';
 
-getPictures().then((photos) => renderPictures(photos)).catch(showDataError);
+let pictures;
 
-document.addEventListener('click', (e) => {
-  const currentPicture = e.target.closest('.picture');
+getData().then((photos) => {
+  pictures = photos;
+  renderPictures(pictures);
+})
+  .catch(showDataError);
+
+document.addEventListener('click', (evt) => {
+  const currentPicture = evt.target.closest('.picture');
 
   if (currentPicture) {
-    e.preventDefault();
-    getPictures().then((photos) => openBigPhotoPopup(photos, currentPicture.dataset.photoId), showDataError);
+    evt.preventDefault();
+    openBigPhotoPopup(pictures, currentPicture.dataset.photoId);
   }
+
 });
 
 setUploadFormSubmit(closeUploadOverlay);
